Removed kronecker graph test
This test allocates lots of data and free it quickly. The algorithm
doesn't deal with this well as it holds all of the data it touched
alive. This means that if it is not able to process data fast enough (as
is the case here), this causes a memory leak.